god damn
none of the serializers worked, I'm sure I'm stupid, but I'm getting annoyed, and my keyboards fear me already.

So I've wasted about 3 hours trying to make either the symfony serializer and the jms serializer, so I'm just using the SimpleXML lib for now, this is the sort of thing that if you show me, in 30 seconds I'll realize what I did wrong, and I'll know it forever.